python_scripts/requirements.txt
altgraph==0.16.1
crowdworks==0.0.1
Django==2.2.2
et-xmlfile==1.0.1
future==0.17.1
image==1.5.27
isodate==0.6.0
jdcal==1.4
macholib==1.11
numpy==1.16.4
openpyxl==2.6.2
pandas==0.24.2
pefile==2019.4.18
Pillow==6.0.0
PyInstaller==3.4
PyMySQL==0.9.3
pyparsing==2.4.0
PySide2==5.12.2
python-dateutil==2.8.0
pytz==2019.1
rdflib==4.2.2
shiboken2==5.12.2
six==1.12.0
SPARQLWrapper==1.8.4
sqlparse==0.3.0
xlrd==1.2.0
